% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{CI_t}
\alias{CI_t}
\title{CI t table}
\usage{
CI_t(x, ci = 0.95)
}
\arguments{
\item{x}{numeric variable}

\item{ci}{CI level}
}
\value{
Returns a tibble with sample size, mean, sd, Margin error, CI_l and CI_U
}
\description{
Computes CI based on t table
}
\examples{
vec <- rnorm(15, mean = 90, sd = 15)
CI_t(vec, ci=0.95)
}
