% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{CI_z}
\alias{CI_z}
\title{CI z table}
\usage{
CI_z(x, ci = 0.95)
}
\arguments{
\item{x}{numeric variable}

\item{ci}{CI level}
}
\value{
Returns a tibble with sample size, mean, sd, Margin error, CI_l and CI_U
}
\description{
Computes CI based on z table
}
\examples{
vec <- rnorm(100, mean = 90, sd = 15)
CI_z(vec, ci=0.95)
}
